hmapi/requirements.txt
alembic==1.7.4
aniso8601==7.0.0
anyio==3.3.3
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
black==21.9b0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.3
distro==1.6.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.17.0
fastapi==0.70.0
filelock==3.3.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
paramiko==2.8.0
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
promise==2.3
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
pywin32==227
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
starlette==0.16.0
texttable==1.6.4
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.8.1
websocket-client==0.59.0
Werkzeug==2.0.1
zipp==3.6.0
